% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/calendarAutoReport.R
\name{calendarAutoReport}
\alias{calendarAutoReport}
\title{Make calendar of report schedule and daily load}
\usage{
calendarAutoReport(runDayOfYear)
}
\arguments{
\item{runDayOfYear}{Numeric vector of yeardays for which reports are to be
run. May well contain repeating values}
}
\value{
a (gg)plot object
}
\description{
Make calendar of report schedule and daily load
}
\examples{
monthlyReport <- seq(from = 1, to = 365, by = 30)
weeklyReport <- seq(from = 1, to = 365, by = 7)
summerWeeklyReport <- seq(from = 130, to = 250, by = 7)
calendarAutoReport(c(monthlyReport, weeklyReport, summerWeeklyReport))
}

