\name{ConvertAPTtoPOS}
\alias{ConvertAPTtoPOS}
\title{ConvertAPTtoPOS}
\usage{
ConvertAPTtoPOS(APTFilePath)
}
\description{
This function will take a .APT file and convert it to a .pos file containing only x, y, z, m information.  File will be saved in same folder as .APT file.}
\examples{

}
