% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raptools_data.R
\docType{data}
\name{data_class}
\alias{data_class}
\title{Simple data set with classifications}
\format{
data frame with 3 columns
\describe{
  \item{ref_class}{The class of the baseline model. Must be a factor}
  \item{new_class}{The class  of the new model. Must be a factor}
  \item{Outcome}{The outcome of interest (Low or High). Must be a factor}
}
}
\usage{
data_class
}
\description{
Example data for use with CI.classNRI
}
\keyword{datasets}
