% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raptools_data.R
\docType{data}
\name{data_risk}
\alias{data_risk}
\title{Simple data set with risk predictions}
\format{
data frame with 3 columns
\describe{
  \item{ref}{The prediction from the baseline model}
  \item{new}{The prediction from the new model}
  \item{outcome}{The outcome of interest (0 or 1)}
}
}
\usage{
data_risk
}
\description{
Example data for use with CI.raplot
}
\keyword{datasets}
