% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConfigTools.R
\name{getConfigTools}
\alias{getConfigTools}
\title{Gets configuration files}
\usage{
getConfigTools(fileName = "rapbaseConfig")
}
\arguments{
\item{fileName}{a string that determines which config file to retrieve.
the default is "rapbaseConfig"}
}
\value{
config file based on the input
}
\description{
Gets configuration files
}
