% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githubApi.R
\name{githubApi}
\alias{githubApi}
\title{Make calls to the github API}
\usage{
githubApi(path, proxyUrl = NULL)
}
\arguments{
\item{path}{Character string with path to the API resource}

\item{proxyUrl}{Character string defining a network proxy in the form
host:port. Default is NULL in which case the API call will not use a proxy}
}
\value{
A list of class githubApi containing the parsed content, API
resource path and the response object
}
\description{
Make calls to the github API
}
