% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installGithubPackage.R
\name{installGithubPackage}
\alias{installGithubPackage}
\title{Install packages from GitHub at Rapporteket}
\usage{
installGithubPackage(
  packageName,
  branchName = "master",
  upgradeDeps = TRUE,
  readConfig = TRUE
)
}
\arguments{
\item{packageName}{String Name of the package}

\item{branchName}{String Name of the branch to use}

\item{upgradeDeps}{Logial if package dependencies should be upgraded}

\item{readConfig}{Set to FALSE to prevent function from reading
configuration. Set TRUE by default. Mainly used for testing purposes.}
}
\value{
story String containing logged entries from the function
}
\description{
Install and report results
}
