% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{prodChecklist}
\alias{prodChecklist}
\title{Checklist when deploying application}
\format{A named list of oks
\describe{
  \item{name}{the text of each check in a PRODUCTION environment}
  \item{values}{dummy ok, not used}
  ...
}}
\usage{
prodChecklist
}
\description{
A dataset containing checklist entries used prior to deploying a registry
app at Rapporteket
}
\keyword{datasets}
