% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raptools.R
\docType{package}
\name{raptools}
\alias{raptools}
\alias{raptools-package}
\title{raptools: A package containing development tools at Rapporteket.}
\description{
Provide functions and resources for development tasks at Rapporteket.
}
\section{raptools functions}{

static_help
}

