\name{read_rrng}
\alias{read_rrng}
\title{read_rrng}
\usage{
read_rrng(RangeFilePath)
}
\description{
Function for reading .rrng files and creating dataframe of values for the variables present.

RangeFilePath is the filepath to the .rrng file that you would like to read.
}
\examples{

}
