\name{readAPTData}
\alias{readAPTData}
\title{readAPTData}
\usage{
readAPTData(DataFilePath)
}
\description{
Reads atom probe data files generated by IVAS/APSuite (in .pos or .APT form) and returns a matrix with column names of variables that are present in file.

Please note that .APT files can contain many variables and will depend on the user's settings in APSuite/IVAS.

DataFilePath is the filepath to the atom probe data file that you would like to read.
}
\examples{

}
