\name{readaptR}
\alias{readaptR}
\title{readaptR}
\usage{
readaptR(posFileName)
}
\description{
Reads .APT files generated by IVAS/APSuite and returns a matrix with column names of variables that are present in file.
Please note that these files can contain many variables and will depend on the user's settings in APSuite/IVAS
posFileName is the filepath to the .APT file that you would like to read.
}
\examples{

}
