% Generated by roxygen2 (4.1.0.9000): do not edit by hand
% Please edit documentation in R/runAutoReport.R
\name{runAutoReport}
\alias{runAutoReport}
\title{Run reports as defined in yaml config}
\usage{
runAutoReport(dayNumber = as.POSIXlt(Sys.Date())$yday + 1)
}
\arguments{
\item{dayNumber}{Integer day of year where January 1st is 1. Defaults to
current day, i.e. as.POSIXlt(Sys.Date())$yday+1}
}
\value{
By itself nothing except whatever the called functions might provide
}
\description{
Usually to be called by a scheduler, e.g. cron. If the proveded day(s) of
year matches the report is run as specified in config
}
\examples{
\dontrun{
runAutoReport()
}
}

