% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_static_help.R
\name{static_help}
\alias{static_help}
\title{Make static html of R doc in package}
\usage{
static_help(pkg, links = tools::findHTMLlinks())
}
\arguments{
\item{pkg}{string providing name of (installed) package containing target
docs}

\item{links}{set NULL to avoid (local) intra-doc linking. Defaults to
tools::findHTMLlinks()}
}
\value{
(a set of) html file(s) in system.file('html', package = pkg)
}
\description{
Make static html of R doc in package
}
\examples{
static_help("raptools", links = NULL)
static_help("raptools")
}
