% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raptools.R
\name{anova_glm}
\alias{anova_glm}
\title{The function anova_glm() returns the Chi^2 and degrees of freedom for each variable & the same was anova.rms() does from lrm() in the rms package.}
\usage{
anova_glm(f)
}
\arguments{
\item{f}{A logistic regression fit created using glm (base package)}
}
\description{
The function anova_glm() returns the Chi^2 and degrees of freedom for each variable & the same was anova.rms() does from lrm() in the rms package.
}
